backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0
python-dotenv==1.0.1
httpx==0.28.1
python-binance==1.0.36
